+++ R/split_df.R
@@ -6,6 +6,8 @@
 #' @param y Name of y variable, default is NULL. The input data will split based on the predictor y, if it is provide.
 #' @param ratio A numeric value, default is 0.7. It indicates the ratio of total rows contained in one split, must less than 1.
 #' @param seed A random seed, default is 618.
+#' @param no_dfs Number of returned data frames. Default is 2.
+#' @param name_dfs Name of returned data frames. If its length is not equal with no_dfs, then the names will seted as 'dX'. Defaults are train and test.
 #'
 #' @return A list of data frames
 #'
@@ -28,7 +30,7 @@
 #'
 #' @import data.table
 #' @export
-split_df = function(dt, y=NULL, ratio=0.7, seed=618) {
+split_df = function(dt, y=NULL, ratio=c(0.7, 0.3), seed=618, no_dfs = 2, name_dfs = c('train', 'test')) {
   ind = NULL
 
   # set dt as data.table
@@ -39,13 +41,19 @@ split_df = function(dt, y=NULL, ratio=0.7, seed=618) {
   # dt = rep_blank_na(dt)
 
   # set ratio range
-  if (!is.numeric(ratio) || length(ratio) >2 || sum(ratio)>1) {
+  if (!is.numeric(ratio) || sum(ratio)>1) {
     warning("Incorrect inputs; ratio must be a numeric that length equal to 1 and less than 1. It was set to 0.7.")
     ratio = c(0.7, 0.3)
   } else {
     ratio_ = 1-sum(ratio)
     if (ratio_ > 0) ratio = c(ratio, ratio_)
   }
+  # no_dfs
+  if (!is.numeric(no_dfs)) {
+    no_dfs = 2
+  } else if (no_dfs > length(ratio)) no_dfs = length(ratio)
+  # name_dfs
+  if (length(name_dfs) != no_dfs) name_dfs = paste0('d',seq_len(no_dfs))
 
   # set seed and partition
   set.seed(seed)
@@ -56,8 +64,7 @@ split_df = function(dt, y=NULL, ratio=0.7, seed=618) {
   }
 
   # random sort
-  rt = list()
-  rt$train = dt[ind == 1,][, ind := NULL]
-  rt$test  = dt[ind == 2,][, ind := NULL]
-  return(rt)
+  lst_dfs = split(dt, by = 'ind', sorted = TRUE, keep.by = FALSE)[seq_len(no_dfs)]
+  names(lst_dfs) = name_dfs
+  return(lst_dfs)
 }